apps/requirement.txt
pandas==1.2.1
streamlit==0.76.0
folium==0.12.1
streamlit-folium-0.1.1